RHOAIENG-79878: Export ModelServingPodSpecOptionsState from @odh-dashboard/hardware-profiles
